+++ test/integration/password_test.rb
@@ -5,27 +5,27 @@ require "test_helper"
 # Tests to assure that passwords can be reset.
 class PasswordTest < ActionDispatch::IntegrationTest
   test "should get forgot your password page" do
-    get new_user_password_path
+    get new_user_password_url
     assert_response :success
   end
 
   test "should request password reset" do
-    post user_password_path, params: {
+    post user_password_url, params: {
       user: { email: users(:admin).email }
     }
     assert_equal I18n.t("devise.passwords.send_instructions"), flash[:notice]
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_url
   end
 
   test "should get reset password page" do
     token = users(:admin).send(:set_reset_password_token)
-    get edit_user_password_path(reset_password_token: token)
+    get edit_user_password_url(reset_password_token: token)
     assert_response :success
   end
 
   test "should reset password" do
     token = users(:admin).send(:set_reset_password_token)
-    patch user_password_path, params: {
+    patch user_password_url, params: {
       user: {
         reset_password_token: token,
         password: "newpassword",